+++ frontend/src/lib/components/bookmark/BookmarkMenuComponent.svelte
@@ -62,5 +62,6 @@
     .lists .bookmark:where(:hover, :focus-visible) {
         background-color: var(--option-active-colour);
         color: var(--on-option-active-colour);
+        outline: 1px solid var(--default-border-color);
     }
 </style>