+++ frontend/src/lib/components/circuit/block/BlockContentComponent.svelte
@@ -6,8 +6,9 @@
     import { getItemsOnPath } from "../../../logic/edition/active_path.svelte";
     import { isSkillItemRevealed } from "../../../logic/circuit/unlocked_skills.svelte";
     import type { SkillItem } from "../../../dto/circuit/edition/skill";
+    import { getAdditionalIcons } from "../../../logic/preferences.svelte";
 
-    let { block }: { block: Block } = $props();
+    let { block, completed }: { block: Block; completed: boolean } = $props();
 
     function isSkillItemVisible(item: SkillItem) {
         return item.column !== null && (!item.hidden || isSkillItemRevealed(item));
@@ -43,7 +44,9 @@
     {#if block.blockType === "skill" && !block.essential}
         <span class="label">Optional</span>
     {/if}
-    <span class="name">{block.name}</span>
+    <span class="name">
+        {block.name}
+    </span>
 </div>
 
 {#if block.blockType === "skill"}
@@ -51,15 +54,15 @@
 {:else if hasEditorRights()}
     <span>{block.items.length} {getLevel().items}</span>
 {:else}
-    {@const completed = getNumCompletedItems("essential")}
-    {@const total = getNumTotalItems("essential")}
+    {@const nrCompleted = getNumCompletedItems("essential")}
+    {@const nrTotal = getNumTotalItems("essential")}
     {@const completedOpt = getNumCompletedItems("optional")}
     {@const totalOpt = getNumTotalItems("optional")}
 
     <div class="completion-counters">
-        {#if total > 0}
+        {#if nrTotal > 0}
             <span>
-                {completed}/{total} completed
+                {nrCompleted}/{nrTotal} completed
             </span>
         {/if}
         {#if block.blockType === "submodule" && totalOpt > 0}
@@ -68,6 +71,10 @@
             </span>
         {/if}
     </div>
+{/if}
+
+{#if completed && getAdditionalIcons()}
+    <span class="checkmark fa-solid fa-check"></span>
 {/if}
 
 <style>
@@ -84,7 +91,7 @@
 
     .label {
         font-style: italic;
-        opacity: 35%;
+        opacity: var(--reduced-opacity);
         margin-top: -0.25em;
     }
 
@@ -95,6 +102,13 @@
 
     .optional-counter {
         font-style: italic;
-        opacity: 35%;
+        opacity: var(--reduced-opacity);
+    }
+
+    .checkmark {
+        position: absolute;
+        color: var(--on-block-task-completed-colour);
+        bottom: 0.7em;
+        right: 1em;
     }
 </style>