+++ frontend/src/lib/components/circuit/block/BookmarkSkillButtonComponent.svelte
@@ -58,7 +58,7 @@
 <style>
     .bookmarks {
         position: absolute;
-        right: -0.5em;
+        right: -0.4em;
         top: -0.5em;
         transform-origin: bottom left;
         /* Must be above blocks */