+++ frontend/src/lib/components/circuit/item/TaskEditComponent.svelte
@@ -126,6 +126,6 @@
 
     .grip {
         cursor: grab;
-        opacity: 0.5;
+        color: var(--drag-icon-color);
     }
 </style>