+++ frontend/src/lib/components/side_controls/bookmarks/BookmarkListComponent.svelte
@@ -268,5 +268,6 @@
     .button:focus-visible,
     .button:hover {
         background: var(--on-glass-surface-active-colour);
+        outline: 1px solid var(--default-border-color);
     }
 </style>