+++ frontend/src/lib/components/side_controls/config/ConfigPanelComponent.svelte
@@ -152,6 +152,7 @@
     .button:focus-visible,
     .button:hover {
         background: var(--on-glass-surface-active-colour);
+        outline: 1px solid var(--default-border-color);
     }
 
     .content {