+++ frontend/src/lib/components/side_controls/paths/PathsPanelComponent.svelte
@@ -132,5 +132,6 @@
     .button:focus-visible,
     .button:hover {
         background: var(--on-glass-surface-active-colour);
+        outline: 1px solid var(--default-border-color);
     }
 </style>