+++ frontend/src/lib/components/side_controls/student_tray/StudentTrayComponent.svelte
@@ -148,6 +148,7 @@
     .button:focus-visible,
     .button:hover {
         background: var(--on-glass-surface-active-colour);
+        outline: 1px solid var(--default-border-color);
     }
 
     .tasks {