+++ frontend/src/lib/components/util/Select.svelte
@@ -271,6 +271,7 @@
     .option:focus-visible {
         background-color: var(--option-active-colour);
         color: var(--on-option-active-colour);
+        outline: 1px solid var(--default-border-color);
     }
 
     .option[aria-selected="false"] .check {