+++ frontend/src/lib/logic/preferences.svelte.ts
@@ -1,9 +1,19 @@
 import type { Preferences } from "../dto/preferences";
-import { lightTheme, systemTheme, type Theme } from "../data/theme";
+import {
+    darkHighContrastTheme,
+    lightHighContrastTheme,
+    lightTheme,
+    systemTheme,
+    type Theme,
+} from "../data/theme";
 import { withCsrf } from "./csrf";
 import { setThemeProperties } from "./theme.svelte";
 
-let preferences: Preferences = $state({ theme: systemTheme, blurBlocks: true });
+let preferences: Preferences = $state({
+    theme: systemTheme,
+    blurBlocks: true,
+    additionalIcons: false,
+});
 
 export async function fetchPreferences() {
     let response = await fetch("/api/person/preferences");
@@ -16,6 +26,10 @@ export function getTheme(): Theme {
 
 export function getBlurBlocks(): boolean {
     return preferences.blurBlocks;
+}
+
+export function getAdditionalIcons(): boolean {
+    return preferences.additionalIcons;
 }
 
 export async function setTheme(theme: Theme) {
@@ -37,4 +51,21 @@ export async function setBlurBlocks(blurBlocksSetting: boolean) {
         }),
     );
     preferences = await response.json();
+}
+
+export async function setAdditionalIcons(additionalIconsSetting: boolean) {
+    let response = await fetch(
+        `/api/person/preferences/additional-icons?additionalIcons=${additionalIconsSetting}`,
+        withCsrf({
+            method: "PATCH",
+        }),
+    );
+    preferences = await response.json();
+}
+
+export function isHighContrastThemeSet(): boolean {
+    return (
+        getTheme().name === lightHighContrastTheme.name ||
+        getTheme().name === darkHighContrastTheme.name
+    );
 }