+++ frontend/src/lib/pages/SettingsPage.svelte
@@ -1,15 +1,26 @@
 <script lang="ts">
     import PageLayout from "./PageLayout.svelte";
     import ThemeSelectComponent from "../components/ThemeSelectComponent.svelte";
-    import { getBlurBlocks, getTheme, setBlurBlocks } from "../logic/preferences.svelte";
-    import ExampleSkillComponent from "../components/circuit/block/ExampleSkillComponent.svelte";
+    import {
+        getAdditionalIcons,
+        getBlurBlocks,
+        getTheme,
+        setAdditionalIcons,
+        setBlurBlocks,
+    } from "../logic/preferences.svelte";
+    import ExampleBlurredSkillComponent from "../components/settings_examples/ExampleBlurredSkillComponent.svelte";
     import { clearLevel, setLevel } from "../logic/circuit/level.svelte";
     import { ModuleLevel } from "../data/level";
+    import ExampleCompletionIcons from "../components/settings_examples/ExampleCompletionIcons.svelte";
 
     clearLevel();
 
     async function updateBlurBlocks(event: Event) {
         await setBlurBlocks((event.target as HTMLInputElement).checked);
+    }
+
+    async function updateAdditionalIcons(event: Event) {
+        await setAdditionalIcons((event.target as HTMLInputElement).checked);
     }
 </script>
 
@@ -25,18 +36,36 @@
         </p>
         <ThemeSelectComponent></ThemeSelectComponent>
 
+        <h2>Additional icons</h2>
+        <p>Configure whether icons indicating the completion of content should be displayed.</p>
+        <label for="completion-icons">
+            <input
+                id="completion-icons"
+                type="checkbox"
+                checked={getAdditionalIcons()}
+                onchange={e => updateAdditionalIcons(e)} />
+
+            Enable completion icons (see example below)
+        </label>
+
+        <ExampleCompletionIcons></ExampleCompletionIcons>
+
         <h2>Unreached content</h2>
         <p>
             Configure whether blocks (e.g., skills and submodules) you have not yet reached should
             be blurred, unless hovered over.
         </p>
-        <div>
-            <input type="checkbox" checked={getBlurBlocks()} onchange={e => updateBlurBlocks(e)} />
+        <label for="block-blurring">
+            <input
+                id="block-blurring"
+                type="checkbox"
+                checked={getBlurBlocks()}
+                onchange={e => updateBlurBlocks(e)} />
 
             Enable block blurring (see example below)
-        </div>
+        </label>
 
-        <ExampleSkillComponent locked={getBlurBlocks()}></ExampleSkillComponent>
+        <ExampleBlurredSkillComponent></ExampleBlurredSkillComponent>
     </div>
 </PageLayout>